4.21-4.22

+++ blogi/controllers/blogs.js
@@ -1,33 +1,22 @@
 const blogsRouter = require("express").Router();
-const jwt = require("jsonwebtoken");
 const Blog = require("../models/blog");
-const User = require("../models/user");
+const tokenExtractor = require("../middlewares/tokenExtractor")
+const userExtractor = require("../middlewares/userExtractor");
 
-blogsRouter.get("/", (request, response) => {
-  Blog.find({})
-    .populate("user", { username: 1, name: 1 })
-    .then((blogs) => {
-      response.json(blogs);
-    });
-});
-
-blogsRouter.post("/", async (request, response, next) => {
+blogsRouter.get("/", async (request, response, next) => {
   try {
+    const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 })
+    response.json(blogs)
+  } catch (error) {
+    next(error)
+  }
+})
+
+
+blogsRouter.post("/", tokenExtractor, userExtractor, async (request, response, next) => {
+  try {
+    const user = request.user;
     const body = request.body;
-    const token = request.token
-    if (!token) {
-      return response.status(401).json({ error: "token missing" })
-    }
-
-    const decodedToken = jwt.verify(token, process.env.SECRET);
-    if (!decodedToken.id) {
-      return response.status(401).json({ error: "token invalid" });
-    }
-
-    const user = await User.findById(decodedToken.id);
-    if (!user) {
-      return response.status(400).json({ error: "UserId missing or not valid" });
-    }
 
     const blog = new Blog({
       title: body.title,
@@ -47,16 +36,32 @@ blogsRouter.post("/", async (request, response, next) => {
   }
 });
 
-blogsRouter.delete("/:id", async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id);
-  response.status(204).end();
+blogsRouter.delete("/:id", tokenExtractor, userExtractor, async (request, response, next) => {
+  try {
+    const user = request.user;
+    const blog = await Blog.findById(request.params.id);
+    if (!blog) {
+      return response.status(404).json({ error: "blog not found" });
+    }
+
+    if (blog.user.toString() === user.id.toString()) {
+      await Blog.findByIdAndDelete(request.params.id);
+      response.status(204).end();
+    } else {
+      return response
+        .status(403)
+        .json({ error: "only the creator can delete this blog" });
+    }
+  } catch (error) {
+    next(error);
+  }
 });
 
 blogsRouter.put("/:id", async (request, response) => {
   const blog = request.body;
   const payload = { likes: blog.likes };
-  await Blog.findByIdAndUpdate(request.params.id, payload, { new: true });
-  response.status(200).end();
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, payload, { new: true });
+  response.status(200).json(updatedBlog);
 });
 
 module.exports = blogsRouter;
